+++ openslides_backend/action/mixins/import_mixins.py
@@ -1,7 +1,7 @@
 import csv
 from enum import Enum
 from time import mktime, strptime, time
-from typing import Any, Callable, Dict, List, Optional, TypedDict
+from typing import Any, Callable, Dict, List, Optional, Set, TypedDict
 
 from ...services.datastore.commands import GetManyRequest
 from ...shared.exceptions import ActionException
@@ -25,6 +25,74 @@ class ImportState(str, Enum):
     GENERATED = "generated"
     REMOVE = "remove"
     ERROR = "error"
+
+
+class ResultType(Enum):
+    """Used by Lookup to differ the possible results in check_duplicate."""
+
+    FOUND_ID = 1
+    FOUND_MORE_IDS = 2
+    NOT_FOUND = 3
+
+
+class Lookup:
+    def __init__(
+        self,
+        datastore: DatastoreService,
+        collection: str,
+        names: List[str],
+        field: str = "name",
+        mapped_fields: List[str] = [],
+    ) -> None:
+        self.datastore = datastore
+        self.collection = collection
+        self.field = field
+        self.name_to_ids: Dict[str, List[Dict[str, Any]]] = {name: [] for name in names}
+        self.id_to_name: Dict[int, str] = {}
+        if "id" not in mapped_fields:
+            mapped_fields.append("id")
+        if field not in mapped_fields:
+            mapped_fields.append(field)
+        if names:
+            for entry in datastore.filter(
+                collection,
+                Or(*[FilterOperator(field, "=", name) for name in names]),
+                mapped_fields,
+                lock_result=False,
+            ).values():
+                self.name_to_ids[entry[field]].append(entry)
+                self.id_to_name[entry["id"]] = entry[field]
+
+    def check_duplicate(self, name: str) -> ResultType:
+        if not self.name_to_ids.get(name):
+            return ResultType.NOT_FOUND
+        elif len(self.name_to_ids[name]) > 1:
+            return ResultType.FOUND_MORE_IDS
+        else:
+            return ResultType.FOUND_ID
+
+    def get_id_by_name(self, name: str) -> Optional[int]:
+        if len(self.name_to_ids[name]) == 1:
+            return self.name_to_ids[name][0]["id"]
+        return None
+
+    def get_name_by_id(self, id_: int) -> Optional[str]:
+        if name := self.id_to_name.get(id_):
+            return name
+        return None
+
+    def read_missing_ids(self, ids: List[int]) -> None:
+        result = self.datastore.get_many(
+            [GetManyRequest(self.collection, ids, [self.field])],
+            lock_result=False,
+            use_changed_models=False,
+        )
+        self.id_to_name.update(
+            {
+                key: value.get(self.field, "")
+                for key, value in result[self.collection].items()
+            }
+        )
 
 
 class ImportMixin(SingularActionMixin):
@@ -102,8 +170,18 @@ class JsonUploadMixin(SingularActionMixin):
     statistics: List[StatisticEntry]
     import_state: ImportState
 
+    """
+    # next fields have to be set in subclass, see committee/json_upload.py.
+    # Implements helpful handling for list fields like admins, organization_tags
+    """
+    import_object_lookup: Lookup  # to be defined and initiated in subclass
+    import_object_name: str  # to be set in subclass for each entry
+    payload_db_field: Dict[str, str]  # mapping payload fieldname to db-id-field
+
     def set_state(self, number_errors: int, number_warnings: int) -> None:
-        # todo: wenn das der import state ist, dann geht das einfacher am ende jeder row. WARNING ghet auch nicht als import state, da row state niemals warning ist oder wir wollen das so
+        """
+        To remove, but is used in some backend imports
+        """
         if number_errors > 0:
             self.import_state = ImportState.ERROR
         elif number_warnings > 0:
@@ -207,70 +285,90 @@ class JsonUploadMixin(SingularActionMixin):
                         )
         super().validate_instance(instance)
 
-
-class ResultType(Enum):
-    """Used by Lookup to differ the possible results in check_duplicate."""
-
-    FOUND_ID = 1
-    FOUND_MORE_IDS = 2
-    NOT_FOUND = 3
-
-
-class Lookup:
-    def __init__(
+    def check_list_field(
         self,
-        datastore: DatastoreService,
-        collection: str,
-        names: List[str],
-        field: str = "name",
-        mapped_fields: List[str] = [],
+        field: str,
+        entry: Dict[str, Any],
+        lookup: Lookup,
+        messages: List[str],
+        not_found_state: ImportState = ImportState.ERROR,
     ) -> None:
-        self.datastore = datastore
-        self.collection = collection
-        self.field = field
-        self.name_to_ids: Dict[str, List[Dict[str, Any]]] = {name: [] for name in names}
-        self.id_to_name: Dict[int, str] = {}
-        if "id" not in mapped_fields:
-            mapped_fields.append("id")
-        if field not in mapped_fields:
-            mapped_fields.append(field)
-        if names:
-            for entry in datastore.filter(
-                collection,
-                Or(*[FilterOperator(field, "=", name) for name in names]),
-                mapped_fields,
-                lock_result=False,
-            ).values():
-                self.name_to_ids[entry[field]].append(entry)
-                self.id_to_name[entry["id"]] = entry[field]
+        if field in entry:
+            # check for parse error
+            if isinstance(entry[field], str):
+                entry[field] = [entry[field]]
 
-    def check_duplicate(self, name: str) -> ResultType:
-        if not self.name_to_ids.get(name):
-            return ResultType.NOT_FOUND
-        elif len(self.name_to_ids[name]) > 1:
-            return ResultType.FOUND_MORE_IDS
-        else:
-            return ResultType.FOUND_ID
+            # found_list and remove_duplicate_list are used to cut duplicates
+            found_for_duplicate_list: List[str] = []
+            remove_duplicate_list: List[str] = []
+            remove_list: List[str] = []
+            not_unique_list: List[str] = []
+            missing_list: List[str] = []
+            db_set_names: Set[str] = set()
 
-    def get_id_by_name(self, name: str) -> Optional[int]:
-        if len(self.name_to_ids[name]) == 1:
-            return self.name_to_ids[name][0]["id"]
-        return None
+            # removed objects
+            if db_field := self.payload_db_field.get(field):
+                db_list_ids = (
+                    self.import_object_lookup.name_to_ids[self.import_object_name][
+                        0
+                    ].get(db_field, [])
+                    or []
+                )
+                db_set_names = set([lookup.id_to_name[id_] for id_ in db_list_ids])
+                new_list_names = set(entry[field])
+                remove_list = list(db_set_names - new_list_names)
+                remove_list.sort()  # necessary for test
 
-    def get_name_by_id(self, id_: int) -> Optional[str]:
-        if name := self.id_to_name.get(id_):
-            return name
-        return None
+            for i, name in enumerate(entry[field]):
+                if name in found_for_duplicate_list:
+                    remove_duplicate_list.append(name)
+                    entry[field][i] = {"value": name, "info": ImportState.WARNING}
+                    continue
 
-    def read_missing_ids(self, ids: List[int]) -> None:
-        result = self.datastore.get_many(
-            [GetManyRequest(self.collection, ids, [self.field])],
-            lock_result=False,
-            use_changed_models=False,
-        )
-        self.id_to_name.update(
-            {
-                key: value.get(self.field, "")
-                for key, value in result[self.collection].items()
-            }
-        )
+                found_for_duplicate_list.append(name)
+                check_duplicate = lookup.check_duplicate(name)
+                if check_duplicate == ResultType.FOUND_ID:
+                    id_ = lookup.get_id_by_name(name)
+                    entry[field][i] = {
+                        "value": name,
+                        # import states signalize a new relation, not the creation of an element
+                        "info": ImportState.DONE
+                        if name in db_set_names
+                        else ImportState.NEW,
+                    }
+                    if id_:
+                        entry[field][i]["id"] = id_
+                elif check_duplicate == ResultType.FOUND_MORE_IDS:
+                    entry[field][i] = {"value": name, "info": ImportState.WARNING}
+                    not_unique_list.append(name)
+                else:
+                    entry[field][i] = {"value": name, "info": not_found_state}
+                    if not_found_state != ImportState.NEW:
+                        missing_list.append(name)
+
+            self.append_message_for_list_fields(
+                not_unique_list,
+                "Not identifiable {field}, because name not unique: [{incorrects}]",
+                field,
+                messages,
+            )
+            self.append_message_for_list_fields(
+                missing_list, "Missing {field}: [{incorrects}]", field, messages
+            )
+            self.append_message_for_list_fields(
+                remove_list, "Removed {field}: [{incorrects}]", field, messages
+            )
+            self.append_message_for_list_fields(
+                remove_duplicate_list,
+                "Removed duplicated {field}: [{incorrects}]",
+                field,
+                messages,
+            )
+
+    def append_message_for_list_fields(
+        self, list_names: List[str], template: str, field: str, messages: List[str]
+    ) -> None:
+        if list_names:
+            list_str = ", ".join(list_names)
+            object = field.replace("_", " ")[:-1] + "(s)"
+            messages.append(template.format(field=object, incorrects=list_str))
